Fix docker setup
- Add dev and prod configs
- Add scripts
- Change folder structure to be conducive to volume mount source code in a container
