+++ package.json
@@ -1,26 +1,28 @@
 {
-  "name": "neotwit",
-  "version": "0.0.0",
+  "name": "maxitwit",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www",
-    "devstart": "nodemon ./bin/www"
+    "start": "node src/bin/www",
+    "devstart": "nodemon src/bin/www",
+    "lint": "eslint --fix .",
+    "docker": "docker-compose up --build"
   },
   "dependencies": {
     "bcrypt": "^5.1.1",
     "connect-flash": "^0.1.1",
     "cookie-parser": "~1.4.4",
-    "debug": "~2.6.9",
     "express": "^4.18.2",
     "express-session": "^1.18.0",
     "express-validator": "^7.0.1",
     "http-errors": "~1.6.3",
-    "morgan": "~1.9.1",
-    "nodemon": "^3.0.3",
     "pug": "^3.0.2",
     "sqlite3": "^5.1.7"
   },
   "devDependencies": {
+    "debug": "~2.6.9",
+    "morgan": "~1.9.1",
+    "nodemon": "^3.0.3",
     "eslint": "^8.56.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.29.1"
