+++ src/views/register.pug
@@ -1,4 +1,4 @@
-extends layout.pug
+extends layout
 
 block content
   h2= title
